Experimental WebSocket Port
This is an experimental port of the Syncplay Protocol to WebSockets.

TLS code has been removed, since that should be handled by an upstream source like Nginx.

DB and Stats code has been removed to cleanup codebase.

Syncplay-proxy will be used to maintain backwards compatibility with the original TCP + STARTTLS protocol.

This code is untested and shouldn't be expected to work at this time.